ArtificialIntelligence/demoapp/spring-app/requirements.txt
boto3==1.28.85
botocore==1.31.85
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
openpyxl==3.1.2
requests==2.31.0
